Major fix: rankings working

+++ monitoring-system/src/main/java/it/uniroma2/sdcc/trafficcontrol/entity/ranking/Rankable.java
@@ -4,7 +4,7 @@ public interface Rankable extends Comparable<Rankable> {
 
     Object getObject();
 
-    long getMeanIntersectionSpeed();
+    int getMeanIntersectionSpeed();
 
     /**
      * Note: We do not defensively copy the object wrapped by the Rankable.  It is passed as is.
